Move Cloudwatch metric code into its own package
